cansat/bridge/requirements.txt
pyserial==3.5
python-socketio[asyncio_client]==5.11.2
requests==2.32.3
